+++ task-sdk/src/airflow/sdk/api/datamodels/_generated.py
@@ -724,6 +724,7 @@ class TITerminalStatePayload(BaseModel):
     state: TerminalStateNonSuccess
     end_date: Annotated[AwareDatetime, Field(title="End Date")]
     rendered_map_index: Annotated[str | None, Field(title="Rendered Map Index")] = None
+    retry_reason: Annotated[str | None, Field(title="Retry Reason")] = None
 
 
 class XComArgBinding(BaseModel):
@@ -829,3 +830,4 @@ class TIRunContext(BaseModel):
     should_retry: Annotated[bool | None, Field(title="Should Retry")] = False
     start_date: Annotated[AwareDatetime | None, Field(title="Start Date")] = None
     arg_bindings: Annotated[list[TaskArgBinding] | None, Field(title="Arg Bindings")] = None
+    multi_team: Annotated[bool | None, Field(title="Multi Team")] = False